Autowire HealthCheck and BardService interfaces
