+++ service/src/test/java/bio/terra/cbas/dao/TestRunDao.java
@@ -6,7 +6,6 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.fail;
 
 import bio.terra.cbas.dao.util.ContainerizedDatabaseTest;
-import bio.terra.cbas.dependencies.bard.BardService;
 import bio.terra.cbas.models.CbasRunSetStatus;
 import bio.terra.cbas.models.CbasRunStatus;
 import bio.terra.cbas.models.GithubMethodVersionDetails;
@@ -22,10 +21,8 @@ import java.util.UUID;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.mock.mockito.MockBean;
 
 class TestRunDao extends ContainerizedDatabaseTest {
-  @MockBean BardService bardService;
   @Autowired RunDao runDao;
   @Autowired RunSetDao runSetDao;
   @Autowired MethodDao methodDao;
