+++ service/src/test/java/bio/terra/cbas/dao/TestRunSetDao.java
@@ -6,7 +6,6 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import bio.terra.cbas.dao.util.ContainerizedDatabaseTest;
-import bio.terra.cbas.dependencies.bard.BardService;
 import bio.terra.cbas.models.CbasRunSetStatus;
 import bio.terra.cbas.models.GithubMethodVersionDetails;
 import bio.terra.cbas.models.Method;
@@ -20,10 +19,9 @@ import java.util.UUID;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.mock.mockito.MockBean;
 
 class TestRunSetDao extends ContainerizedDatabaseTest {
-  @MockBean BardService bardService;
+
   @Autowired RunSetDao runSetDao;
   @Autowired MethodDao methodDao;
   @Autowired MethodVersionDao methodVersionDao;
